Add files via upload
In this project, the main objective is to create a deep learning model for the classification of skin images into two categories: benign (a type of skin cancer) and malignant. To achieve this, the project will employ the EfficientNet Lite model, which is a variant of the EfficientNet architecture designed for efficient and lightweight use in mobile and resource-constrained environments.